build 1.0
In this build, hitting the Enter key will automatically convert the text. It could be done automatically after every single key press, but there is a chance that it will consume more performance as it will call the conversion method every time any key is pressed.

Besides that, refactored the name of the fields. So, instead of "label1", now it is possible to locate the item as "TopLabel", for example. This brings a better readability for the code.

Also, includes some changes in design.